+++ src/main/scala/conversion/PartiallyOrderedNode.scala
@@ -1,0 +1,12 @@
+package com.klout.scoozie
+package conversion
+
+case class PartiallyOrderedNode(
+    node:         GraphNode,
+    partialOrder: Int)
+
+case object PartiallyOrderedNode {
+    def lt(x: PartiallyOrderedNode, y: PartiallyOrderedNode): Boolean = {
+        x.partialOrder < y.partialOrder || (x.partialOrder == y.partialOrder && x.node.name < y.node.name)
+    }
+}
