+++ src/main/scala/conversion/RefWrap.scala
@@ -1,0 +1,10 @@
+package com.klout.scoozie
+package conversion
+
+case class RefWrap[T <: AnyRef](value: T) {
+    override def equals(other: Any): Boolean = other match {
+        case ref: RefWrap[_] => ref.value eq value
+        case _               => false
+    }
+}
+
