+++ src/main/scala/dsl/Decision.scala
@@ -1,0 +1,11 @@
+package com.klout.scoozie
+package dsl
+
+case class Decision(predicates: List[(String, Predicate)]) {
+    def dependsOn(dep1: Dependency, deps: Dependency*): DecisionNode = DecisionNode(this, Set(dep1) ++ deps)
+    def dependsOn(deps: Seq[Dependency]): DecisionNode = DecisionNode(this, deps.toSet)
+}
+
+object Decision {
+    def apply(pair1: (String, Predicate), pairs: (String, Predicate)*): Decision = Decision(pair1 :: pairs.toList)
+}
