+++ src/main/scala/dsl/SugarNode.scala
@@ -1,0 +1,4 @@
+package com.klout.scoozie
+package dsl
+
+case class SugarNode(work: Work, dependency: DecisionDependency, previousSugarNode: Option[SugarNode] = None)
