+++ src/main/scala/runner/OozieConfig.scala
@@ -1,0 +1,4 @@
+package com.klout.scoozie
+package runner
+
+case class OozieConfig(oozieUrl: String, properties: Map[String, String])
