+++ src/main/scala/runner/OozieError.scala
@@ -1,0 +1,4 @@
+package com.klout.scoozie
+package runner
+
+case class OozieError(jobId: String, jobLog: String, consoleUrl: String)
