+++ src/main/scala/runner/OozieSuccess.scala
@@ -1,0 +1,4 @@
+package com.klout.scoozie
+package runner
+
+case class OozieSuccess(jobId: String)
