+++ src/main/scala/runner/XmlPostProcessing.scala
@@ -1,0 +1,12 @@
+package com.klout.scoozie
+package runner
+
+case class XmlPostProcessing(
+    substitutions: Map[String, String])
+
+object XmlPostProcessing {
+    val Default = XmlPostProcessing(
+        substitutions = Map(
+            "&quot;" -> "\""))
+}
+
