--- src/main/scala/scoozie.scala
@@ -1,11 +1,0 @@
-/**
- * Copyright (C) 2013 Klout Inc. <http://www.klout.com>
- */
-
-package com.klout.scoozie
-
-object `package` {
-
-    def ???[T] = sys.error("not implemented yet!")
-
-}
