+++ src/main/scala/verification/ForkThread.scala
@@ -1,0 +1,7 @@
+package com.klout.scoozie
+package verification
+
+import com.klout.scoozie.conversion.GraphNode
+
+//path is the child of the fork that this thread lies on
+case class ForkThread(fork: GraphNode, path: GraphNode)
