+++ src/main/scala/verification/VerificationNode.scala
@@ -1,0 +1,6 @@
+package com.klout.scoozie
+package verification
+
+import com.klout.scoozie.conversion.GraphNode
+
+case class VerificationNode(graphNode: GraphNode, parentThreads: Seq[ForkThread])
