+++ src/test/scala/ConversionSpec.scala
